requirements.txt
recipe_scrapers
flask
flask-cors
pillow
requests
python-dotenv
